requirements.txt
requests>=2.31.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
icalendar>=5.0.11
python-dotenv>=1.0.0
